Support mountpoints anywhere in the root filesystem
You can now mount HDD, USB stick, NAS, ... anywhere in the
filesystem and not only under /media and /tmp.
The filesystem with the image files will be moved to new
rootfs. All others will be unmounted during flash.
(This is version 2 of the patch as first version was reverted
this errors)